publish vivek/stats-page along with basic changes
